refactor(points): encapsula lógica de puntos en el modelo y evita condiciones de carrera
Las funciones de sumar y restar puntos se han movido del view al modelo para mejorar el encapsulamiento y evitar condiciones de carrera usando transacciones. También se ha eliminado un serializer de reserva duplicado.
